# mofumofu-ui/src/lib/api/swagger.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "mofumofu-backend",
		"description": "A soft and simple blogging platform for everyone.",
		"contact": { "name": "Levi Lim", "email": "shiueo.csh@gmail.com" },
		"license": { "name": "" },
		"version": "0.6.9"
	},
	"paths": {
...
		"/v0/like": {
			"post": {
				"tags": ["Like"],
				"operationId": "create_like",
				"requestBody": {
					"content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateLikeRequest" } } },
					"required": true
				},
				"responses": {
					"204": { "description": "Post liked successfully" },
					"400": { "description": "Invalid input or already liked" },
					"404": { "description": "Post not found" },
					"500": { "description": "Internal Server Error" }
				},
				"security": [{ "bearer_auth": [] }]
			},
			"delete": {
				"tags": ["Like"],
				"operationId": "delete_like",
				"requestBody": {
					"content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteLikeRequest" } } },
					"required": true
				},
				"responses": {
					"204": { "description": "Post unliked successfully" },
					"400": { "description": "Invalid input or like not found" },
					"404": { "description": "Post not found" },
					"500": { "description": "Internal Server Error" }
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
...
		"/v0/post": {
			"put": {
				"tags": ["Post"],
				"operationId": "update_post",
				"requestBody": {
					"content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdatePostRequest" } } },
					"required": true
				},
				"responses": {
					"204": { "description": "Post updated successfully" },
					"400": { "description": "Invalid input" },
					"401": { "description": "Unauthorized or email not verified" },
					"404": { "description": "Post not found" },
					"500": { "description": "Internal Server Error" }
				},
				"security": [{ "bearer_auth": [] }]
			},
			"post": {
				"tags": ["Post"],
				"operationId": "create_post",
				"requestBody": {
					"content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatePostRequest" } } },
					"required": true
				},
				"responses": {
					"201": {
						"description": "Post created successfully",
						"content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatePostResponse" } } }
					},
					"400": { "description": "Invalid input" },
					"401": { "description": "Unauthorized or email not verified" },
					"500": { "description": "Internal Server Error" }
				},
				"security": [{ "bearer_auth": [] }]
			},
			"delete": {
				"tags": ["Post"],
				"operationId": "delete_post",
				"requestBody": {
					"content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeletePostRequest" } } },
					"required": true
				},
				"responses": {
					"204": { "description": "Post deleted successfully" },
					"400": { "description": "Invalid input" },
					"401": { "description": "Unauthorized or email not verified" },
					"404": { "description": "Post not found" },
					"500": { "description": "Internal Server Error" }
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"CreateLikeRequest": {
				"type": "object",
				"required": ["post_id"],
				"properties": { "post_id": { "type": "string", "format": "uuid" } }
			},
			"CreatePostRequest": {
				"type": "object",
				"required": ["title", "content", "slug"],
				"properties": {
					"content": { "type": "string" },
					"hashtags": { "type": ["array", "null"], "items": { "type": "string" } },
					"slug": { "type": "string" },
					"summary": { "type": ["string", "null"] },
					"title": { "type": "string" }
				}
			},
			"CreatePostResponse": {
				"type": "object",
				"required": ["post_id"],
				"properties": { "post_id": { "type": "string", "format": "uuid" } }
			},
...
			"DeleteLikeRequest": {
				"type": "object",
				"required": ["post_id"],
				"properties": { "post_id": { "type": "string", "format": "uuid" } }
			},
			"DeletePostRequest": { "type": "object", "required": ["slug"], "properties": { "slug": { "type": "string" } } },
...
			"UpdatePostRequest": {
				"type": "object",
				"required": ["post_id"],
				"properties": {
					"content": { "type": ["string", "null"] },
					"hashtags": { "type": ["array", "null"], "items": { "type": "string" } },
					"new_slug": { "type": ["string", "null"] },
					"post_id": { "type": "string", "format": "uuid" },
					"summary": { "type": ["string", "null"] },
					"title": { "type": ["string", "null"] }
				}
			},
...
		},
		"securitySchemes": {
			"anonymous_id_cookie": { "type": "apiKey", "in": "cookie", "name": "anonymous_user_id" },
			"bearer_auth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
			"refresh_token_cookie": { "type": "apiKey", "in": "cookie", "name": "refresh_token" }
		}
	},
...
}